Backend/requirements.txt
Flask
torch
torchvision
numpy
pillow
scikit-image
flask-cors
werkzeug
gunicorn
